pyproject.toml
[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 120

# Enable following rules (see this section : https://github.com/charliermarsh/ruff#supported-rules)
select = ["E", "F", "I", "N", "C", "T"]
ignore = []

# Exclude a variety of commonly ignored directories.
exclude = [
    "examples.py",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

fix = true # Always autofix

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.11.
target-version = "py311"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"main.py" = ["C901"]


[tool.poetry]
name = "SNPrimer"
version = "0.2.0"
description = "Small python library to search snp in primer by position or by sequence."
authors = ["Benoitdw <bw@oncodna.com>"]
readme = "README.md"
repository = "https://github.com/Benoitdw/SNPrimer"


[tool.poetry.dependencies]
python = "^3.11"
myvariant = "^1.0.0"
gget = "^0.3.11"
click = "^8.1.3"
pyfaidx = "^0.7.1"
dataclasses-json = "^0.5.7"
primer3-py = "^0.6.1"

[tool.poetry.group.dev]

[tool.poetry.group.dev.dependencies]
bump2version = "^1.0.1"
pre-commit = "^2.20.0"
pyfaidx = "^0.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
